+++ CHANGELOG.md
@@ -31,7 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-[fj-doc-maven-plugin] m2e lifecycleMappingMetadata (xml is not valid) # 194  
+- [fj-doc-maven-plugin] m2e lifecycleMappingMetadata (xml is not valid) # 194  
 
 ### Changed