+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- [fj-doc-maven-plugin] goal init, flavour quarkus-3, default quarkus version 3.14.3
 - [fj-doc-playground-quarkus] doc project init, default venus version 8.8.5
 - [fj-doc-playground-quarkus] quarkus version set to 3.14.3